% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{package}
\name{assign2landmarks}
\alias{assign2landmarks}
\alias{assign2landmarks-package}
\title{assign cells/samples to the closest package}
\usage{
assign2landmarks(ForeGround, topLandmarks)
}
\arguments{
\item{ForeGround}{matrix or data frame of Foreground values}

\item{topLandmarks}{output from compute_landmarks}
}
\description{
assign cells/samples to the closest package
}
