% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{computeLandmarks}
\alias{computeLandmarks}
\title{compute landmarks}
\usage{
computeLandmarks(ForeGround, BackGround = NULL, weights = NULL,
  nCluster = 2, lambda = 0.1, nTop = 5000)
}
\arguments{
\item{ForeGround}{sparse matrix of Foreground values}

\item{BackGround}{sparse matrix of BackGround values, one of BackGround or weights must be provided}

\item{weights}{a vector of weights to use for clustering, one of BackGround or weights must be provided}

\item{nCluster}{number of clusters (default = 2)}

\item{lambda}{weighting parameter (default = 0.1)}

\item{nTop}{number of top clusters (default = 5000)}
}
\description{
compute landmarks
}
