% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_background}
\alias{filter_background}
\title{filter_background}
\usage{
filter_background(ForeGround, BackGround, median_bg_thresh = 2)
}
\arguments{
\item{ForeGround}{matrix or data frame of Foreground values}

\item{BackGround}{matrix or data frame of BackGround values}

\item{thresh}{threshold of the median number of reads in background}
}
\value{
filtered ForeGround and BackGround
}
\description{
filter_background
}
