% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filter_samples}
\alias{filter_samples}
\title{filter_samples}
\usage{
filter_samples(ForeGround, BackGround, readsFGthresh = NULL)
}
\arguments{
\item{ForeGround}{matrix or data frame of ForeGround values}

\item{BackGround}{matrix or data frame of BackGround values}

\item{readsFGthresh}{threshold for the total reads per cell in ForeGround. Default is min(500, number of peaks/50)}
}
\value{
filtered ForeGround and BackGround
}
\description{
filter_samples
}
