% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{filterSamples}
\alias{filterSamples}
\title{filter samples by comparing foreground against background}
\usage{
filterSamples(ForeGround, BackGround, readsFGthresh = NULL)
}
\arguments{
\item{ForeGround}{sparse matrix of ForeGround values}

\item{BackGround}{sparse of BackGround values}

\item{readsFGthresh}{threshold for the total reads per cell in ForeGround. Default is min(500, number of peaks/50)}
}
\value{
filtered ForeGround and BackGround
}
\description{
filter samples by comparing foreground against background
}
