% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{getBackground}
\alias{getBackground}
\title{compute background counts matrix}
\usage{
getBackground(bamfiles, peaks, upstream = 5e+05, downstream = 5e+05,
  byReadGroup = FALSE, VERBOSE = FALSE, PAIRED = FALSE)
}
\arguments{
\item{bamfiles}{a vector of filenames of input bamfiles}

\item{peaks}{a bed15 format file returned from select peaks}

\item{upstream}{number of bases upstream of peak to consider for computing background}

\item{downstream}{number of bases downstream of peak to consider for computing background}
}
\value{
a matrix of chrom, start, end of peaks followed by background counts for each bam file in bamfiles
}
\description{
compute background counts matrix
}
