% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{getCountsMatrix2}
\alias{getCountsMatrix2}
\title{get counts matrix}
\usage{
getCountsMatrix2(bamfiles, peaks, PAIRED = FALSE, byReadGroup = FALSE,
  VERBOSE = FALSE)
}
\arguments{
\item{bamfiles}{a vector of filenames of input bam files}

\item{peaks}{a bed15 format file returned from select_peaks}

\item{byReadGroup}{boolean variable indicating whether or not individual experiments are separated by read group}

\item{VERBOSE}{a boolean variable to indicate whether to print update messages}

\item{paired}{a boolean variable indicating whether the experiment was paired end or not}
}
\value{
a matrix of chrom, start, end of peaks followed by counts of each bam file in bamfiles
}
\description{
Obtain counts matrix from bamfiles and peaks
}
\keyword{counts}
\keyword{peaks}
