% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plotGapStat}
\alias{plotGapStat}
\title{plotGapStat}
\usage{
plotGapStat(GapStat, nClusters, main, fold = 1, cex = 0.3, lty = 2)
}
\arguments{
\item{GapStat}{output of getGapStat}

\item{nClusters}{vector of clusters to try, default: 1:10}

\item{main}{title of plot}
}
\value{
nothing, just plotting
}
\description{
plotGapStat
}
