% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scABC.R
\name{scABC}
\alias{scABC}
\title{scABC}
\usage{
scABC(bamfiles, peakfile, PLOT = TRUE, QUIET = TRUE, nClusters = c(),
  pValThresh = 2, nreadsThresh = 2, ncellsThresh = 10,
  readsFGthresh = NULL, lambda = 1, nTop = 2000, nPerm = 20,
  nTopGapStat = 5000, maxiter = 1000, thresMLE = 10^-3,
  thresMAP = 10^-5)
}
\arguments{
\item{bamfiles}{location of bam files}

\item{peakfile}{location of peak file}

\item{PLOT}{boolean variable to make plots, default: TRUE}

\item{QUIET}{boolean variable for verbosity, default: TRUE}

\item{nClusters}{number of clusters, if not set manually then the number of clusters is determined by gap stat analysis}

\item{pValThresh}{the -log10(p) threshhold on the MACS2 computed peak p value, default: 2 (p < 0.01)}

\item{nreadsThresh}{threshold on the minimum number of reads in ncellsThresh to require when filtering cells, default: 2}

\item{ncellsThresh}{threshold on the minimum number of cells with nreadsThresh to require when filtering cell, default: 10}

\item{readsFGthresh}{threshold on the number of reads falling in peaks to filter cells by, default: min(500, nrow(ForeGround)/50)}

\item{lambda}{weighting parameter for clustering, default: 1}

\item{nTop}{the number of top peaks to use when forming the landmarks for each cluster, default: 2000}

\item{nPerm}{the number of permutation to use when selecting peaks by the gap statistic, default: 20}

\item{nTopGapStat}{the number of top peaks to use when computing the gap statistic, default: 5000}

\item{maxiter}{maximum number of iterations when computing beta, default: 1000}

\item{thresMLE}{numerical threshhold for convergence of beta MLE, default: 10^-3}

\item{thresMAP}{numerical threshhold for convergence of MAP, default: 10^-5}
}
\value{
a list of the processed ForeGround, peaks, number of clusters, cluster assignments, landmarks, and peak p values
}
\description{
scABC
}
\examples{
scABCcluster = scABC(bamfiles, peakfile, nClusters = 1:5) 
}
\author{
Timothy Daley \email{tdaley@stanford.edu}, Zhixiang Lin \email{zl235@stanford.edu},
        Mahdi Zamanighomi \email{mzamani@stanford.edu}
}
