% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{select_peaks}
\alias{select_peaks}
\title{select peaks}
\usage{
select_peaks(filename, thresh = 2)
}
\arguments{
\item{filename}{of a bed12+3 gapped peaks file obtain from peaking calling using MACS2}
}
\value{
significant peaks obtained by filtering by p-value
}
\description{
select peaks
}
\keyword{peaks}
