% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{computeLandmarks}
\alias{computeLandmarks}
\title{compute landmarks}
\usage{
computeLandmarks(ForeGround, BackGround, nCluster = 2, lambda = 0.1,
  nTop = 2000)
}
\arguments{
\item{ForeGround}{matrix or data frame of Foreground values}

\item{BackGround}{matrix or data frame of BackGround values}

\item{nCluster}{number of clusters (default = 2)}

\item{lambda}{weighting parameter (default = 0.1)}

\item{nTop}{number of top clusters}
}
\description{
compute landmarks
}
