% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{package}
\name{getClusterSpecificPvalue}
\alias{getClusterSpecificPvalue}
\alias{getClusterSpecificPvalue-package}
\title{get cluster specific pvals for peaks}
\usage{
getClusterSpecificPvalue(ForeGround, cluster_assignments, background_medians,
  landmark = NULL, maxiter = 1000, thresMLE = 10^-3, thresMAP = 10^-5,
  quiet = TRUE)
}
\arguments{
\item{landmark}{optional. If landmark is provided, we only do hypothesis testing on the union of landmark peaks}

\item{data}{matrix of peaks by cells counts}

\item{cluster}{matrix of cluster indicator membership}

\item{background_median}{median background values for each cell}
}
\description{
get cluster specific pvals for peaks
}
