% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{getCountsMatrix}
\alias{getCountsMatrix}
\title{get_counts_matrix}
\usage{
getCountsMatrix(bamfiles, peaks)
}
\arguments{
\item{bamfiles}{a vector of filenames of input bam files}

\item{peaks}{a bed15 format file returned from select_peaks}
}
\value{
a matrix of chrom, start, end of peaks followed by counts of each bam file in bamfiles
}
\description{
get_counts_matrix
}
\keyword{counts}
\keyword{peaks}
