% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{getGapStat}
\alias{getGapStat}
\title{get gap statistic and choose best cluster}
\usage{
getGapStat(ForeGround, BackGroundMedian, nClusters = 1:10, nPerm = 20,
  nTop = 5000, quiet = FALSE)
}
\arguments{
\item{ForeGround}{sparse matrix ForeGround values}

\item{BackGroundMedian}{median of background values to compute weighting}

\item{nClusters}{vector of clusters to try, default: 1:10}

\item{nPerm}{number of permutation to perform, default: 10}

\item{nTop}{the number of top features to use in the clustering}
}
\value{
list of gap stat stuff
}
\description{
get gap statistic and choose best cluster
}
