% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scABC.R
\name{scABC}
\alias{scABC}
\title{scABC}
\usage{
scABC(bamfiles, peakfile, PLOT = TRUE, QUIET = TRUE, nClusters = c(),
  pValThresh = 2, nreadsThresh = 2, ncellsThresh = 10,
  medianBGthresh = 2, lambda = 1, nTop = 2000, nPerm = 20,
  maxiter = 1000, thresMLE = 10^-3, thresMAP = 10^-5)
}
\arguments{
\item{bamfiles}{location of bam files}

\item{peakfile}{location of peak file}

\item{nClusters}{number of clusters, if not set manually then the number of clusters is determined by gap stat analysis}

\item{plot}{boolean variable to make plots}
}
\description{
scABC
}
\examples{
bamfiles and peakfile contain the location of the bams and peaks (in bed format)
         scABCcluster = scABC(bamfiles, peakfile, nClusters = 1:5) chooses the optimal 
         cluster from 1 to 5 and returns the filtered data and peak along with cluster
         assignments, representative landmarks, and peak p-values
}
\author{
Timothy Daley \email{tdaley@stanford.edu}, Zhixiang Lin \email{zl235@stanford.edu},
        Mahdi Zamanighomi \email{mzamani@stanford.edu}
}
